Extract out webhook cluster test from workflow
